refactor(codegen): Rename package from codgen to codegen for clarity
