--- internal/codgen/codgen.go
@@ -1,1 +1,0 @@
-package codgen
